profhome-it-bots/requirements.txt
pyTelegramBotAPI==4.12.0
pyyaml==6.0.1